Fixed incorrect ilvl calculation on 2-handed weapon setups
The original code from Dreamstate83 does not calculate 2-handed weapon
setups in 7.2 correctly. Originally if a 2H weapon was detected, we
would take one away from the amount of slots being calculated in the
average (ie. there are 16 slots that you can equip an item to that
affects your ilvl. If a 2H it would be counted as 15). This lead to
small variences in how Blizzard count ilvl's compared to this addon. It
has now been fixed to always count a 2H in both MH and OH slots, for
both ilvl and slot count. This appears to be how Blizzard counts ilvls