Removed LibItemUpgradeInfo-1.0 library references
LibItemUpgradeInfo-1.0 was used to calculate BoA/heirloom equipment
upgrade ilvl values. It seems like Blizzard has made this information
available easily in the API when scanning and item, just as you would
with a normal item. Using this library was causing variences in the ilvl
calculations that didn't quite make sense.
I have left the library intact with the package to ensure nothing breaks
in this release. Will continue pulling apart this code and will aim to
have it completely removed in the future