Fixed addon namespace conflict and staff ilvl calculations
The addon Oilvl appears to have a global function using the same
namespace as TIL. I have renamed the function OTgathertil() to
TILOTgathertil() - may need to come up with better naming convention,
but this fixes the issue for now.

Readded LibItemUpgradeInfo library. It appears that characters with a
staff weapon need this library for the ilvl to return correctly. I need
to look into LibItemUpgradeInfo more to understand exactly why this is
the case